requirements.txt
flask==3.0.3
flask-socketio==5.3.6
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
eventlet==0.35.2
PyJWT==2.8.0
redis==5.0.4
werkzeug==3.0.3
gunicorn==22.0.0
psycopg2-binary==2.9.9
